+++ src/main/java/com/loopmarket/domain/chat/entity/ChatRoomEntity.java
@@ -40,7 +40,7 @@ public class ChatRoomEntity {
     private boolean user2Leaved = false;
     
     @Column(name = "product_id", nullable = false)
-    private Integer productId; // 어떤 상품에 대한 채팅방인지
+    private Long productId; // 어떤 상품에 대한 채팅방인지
     
     @Column(name = "created_at", updatable = false)
     private LocalDateTime createdAt;
