+++ src/main/java/com/loopmarket/domain/chat/repository/ChatRoomRepository.java
@@ -16,7 +16,7 @@ public interface ChatRoomRepository extends JpaRepository<ChatRoomEntity, Long> 
      * productId까지 포함해서 중복 방지
      * - 두 사용자의 조합으로 유일한 채팅방 존재
      */
-	Optional<ChatRoomEntity> findByUser1IdAndUser2IdAndProductId(Integer user1, Integer user2, Integer productId);
+	Optional<ChatRoomEntity> findByUser1IdAndUser2IdAndProductId(Integer user1, Integer user2, Long productId);
 
 
     /**
